Train a model using Cloud ML

Upload a TensorFlow application to Google Cloud, and use that application to train a model.

cloudml_train(file = "train.R", scale_tier = c("basic", "basic-gpu",
  "basic-tpu"), flags = NULL, cloudml = NULL, gcloud = NULL,
  collect = "ask")

Arguments

file
File to be used as entrypoint for training.

scale_tier
Machine type for training. "basic" provides a single worker instance suitable for learning how to use Cloud ML Engine, and for experimenting with new models using small datasets; "basic-gpu" provides a single worker instance with a GPU. "basic-tpu" provides a single worker instance with a TPU.

flags
Named list with flag values (see flags()) or path to YAML file containing flag values.

cloudml
A list, YAML or JSON configuration file as described https://cloud.google.com/ml-engine/reference/rest/v1/projects.jobs.

gcloud
A list or YAML file with optional 'account' or 'project' fields used to configure the GCloud environemnt.

collect
Collect job when training is completed (blocks waiting for the job to complete).

See also

job_status(), job_collect(), job_cancel()